test_flux

+++ ArrayGener2.0/get_coord.m
@@ -2,7 +2,7 @@ function [v1,v2,varargout]=get_coord(X,Mesh)
 % input index of nodes and mesh structure, return the coordinates of input
 % nodes
 nout = max(nargout,1);
-x_B = Mesh.ind2sub(X,:); % xt
+x_B = double(Mesh.ind2sub(X,:)); % xt
 n = length(Mesh.discr_bnd(:,1));
 coord = zeros(n,length(X));
